// src/api/model/assigneeDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */
import type { AssigneeDtoDeletedAt } from './assigneeDtoDeletedAt';
import type { AssigneeDtoDeletedBy } from './assigneeDtoDeletedBy';
import type { AssigneeDtoIcon } from './assigneeDtoIcon';

export interface AssigneeDto {
  createdAt: string;
  createdBy: number;
  updatedAt: string;
  updatedBy: number;
  /** @nullable */
  deletedAt: AssigneeDtoDeletedAt;
  /** @nullable */
  deletedBy: AssigneeDtoDeletedBy;
  id: number;
  name: string;
  color: string;
  /** @nullable */
  icon: AssigneeDtoIcon;
  workspaceId: number;
}